class based top profile container should be ready minus pictures upload
