+++ components/ProfileTopContainer.js
@@ -1,19 +1,24 @@
 import React, {useState} from "react";
 import { View, Text, Image, TouchableOpacity } from "react-native";
-import { StatusBar } from "expo-status-bar";
+import { connect } from "react-redux";
 
 import {
   Colors,
-}from "./styles";
+} from './styles';
 
-export default function ProfileTopContainer({navigation, route}) {
-  
-  // const {fullName, email, id} = route.params;
-  const [profileData, setProfileData] = useState(fullName="Samira Bahram", email="samirabahrampour@gmail.com")
+class ProfileTopContainer extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      fullName: '',
+      email: '',
+    }
+  }
 
-  return (
+  render() {
+    return(
 
-    <View 
+      <View 
       style={{
         flex: 1, 
         justifyContent: 'space-between', 
@@ -24,7 +29,7 @@ export default function ProfileTopContainer({navigation, route}) {
         paddingRight: 15
       }}>
 
-      <Image source={require('../assets/images/foxicon.png')} 
+      <Image source={require('../assets/images/foxicon.png')}
         style={{
           width: 70, 
           height: 70,
@@ -44,14 +49,14 @@ export default function ProfileTopContainer({navigation, route}) {
             paddingRight: 20
           }}>
               
-          <Text style={{fontSize:15}}>{fullName}{"\n"}</Text>
-          <Text style={{fontSize:12}}>{email}{"\n"}</Text>
+          <Text style={{fontSize:15}}>{this.state.fullName}{"\n"}</Text>
+          <Text style={{fontSize:12}}>{this.state.email}{"\n"}</Text>
 
         </Text>
 
           <TouchableOpacity 
             activeOpacity={0.5}
-            onPress = {() => navigation.navigate('ToBeDefined')}
+            onPress = {() => this.props.navigation.navigate('ToBeDefined')}
           >
               
             <Image 
@@ -68,6 +73,22 @@ export default function ProfileTopContainer({navigation, route}) {
 
     </View>
 
-  )
+    );
+  }
 }
 
+const mapStateToProps = state => {
+  const  {fullName} = state;
+  const  {email} = state;
+
+  console.log('user data mapStateToProps ... state:' + JSON.stringify(state));
+
+  return {fullName, email};
+};
+
+export default connect(
+  mapStateToProps,
+)(ProfileTopContainer);
+
+// export default connect(
+// )(ProfileTopContainer);
